+++ src/main/java/org/loverde/geographiccoordinate/compass/CompassDirection32.java
@@ -40,6 +40,7 @@ import java.util.Map;
 import org.loverde.geographiccoordinate.internal.EnumHelper;
 
 import static java.math.BigDecimal.ZERO;
+import static org.loverde.geographiccoordinate.internal.Objects.failIf;
 
 
 /**
@@ -175,9 +176,7 @@ public enum CompassDirection32 implements CompassDirection {
         final CompassDirection32[] values;
         CompassDirection32 dir;
 
-        if (bearing.compareTo(ZERO) < 0 || bearing.compareTo(BD360) > 0) {
-            throw new IllegalArgumentException(String.format("Bearing %s is not in range [0, 360]", bearing.toPlainString()));
-        }
+        failIf(bearing.compareTo(ZERO) < 0 || bearing.compareTo(BD360) > 0, () -> "Bearing %s is not in range [0, 360]".formatted(bearing.toPlainString()));
 
         values = values();
         newBearing = (bearing.compareTo(BD360) == 0 ? ZERO : bearing).setScale(2, RoundingMode.HALF_UP);